# Conflicts:
#	app/Http/Controllers/PaperController.php
#	app/ViewBlock/AdminGlobalVar.php
#	resources/views/adminhtml/templates/papers/edit.blade.php

+++ app/Http/Controllers/WriterController.php
@@ -4,24 +4,30 @@ namespace App\Http\Controllers;
 
 use App\Helper\ImageUpload;
 use App\Models\Writer;
+use App\Models\WriterInterface;
 use Illuminate\Http\Request;
 use Carbon\Carbon;
 
-class WriterController extends Controller
+class WriterController extends Controller implements WriterControllerInterface
 {
     use ImageUpload;
+
     protected $request;
     protected $writer;
 
     public function __construct(
         Request $request,
         Writer $writer
-    ) {
+    )
+    {
         $this->request = $request;
         $this->writer = $writer;
     }
 
-    public function listOfWriter()
+    /**
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     */
+    public function listWriter()
     {
         if (view()->exists('adminhtml.templates.writer.list')) {
             $all_writer = Writer::paginate(8);
@@ -43,7 +49,7 @@ class WriterController extends Controller
     public function insertWriter()
     {
         $request = $this->request;
-        if ($file = $request->__get("image_post")){
+        if ($file = $request->__get("image_post")) {
             $image_upload_path = $this->uploadImage($file, "public/images/writer", "images/resize/writer");
         }
 
@@ -57,13 +63,13 @@ class WriterController extends Controller
             "name_alias" => $request->__get("alias"),
             "active" => $request->__get("active") ?: true,
             "date_of_birth" => Carbon::createFromFormat('Y-m-d', $request->__get("date_of_birth")), // date('Y-m-d H:i:s', strtotime($request->__get("date_of_birth")))
-            "good" => $request->__get("good") ?: null
+            "rating" => $request->__get("good") ?: null
         ]);
 
         $result = $writer->save();
         if ($result) {
             return redirect()->back()->with("success", "created new writer");
-        }else {
+        } else {
             return redirect()->back()->with("error", "create fail!!, please try again.");
         }
     }
@@ -71,20 +77,19 @@ class WriterController extends Controller
     public function deleteWriter()
     {
         try {
-            $request = $this->request;
-        if ($writer_id = $request->__get("writer_id")) {
-            $writer = $this->writer->find($writer_id);
-            if ($writer && $writer->id) {
-                // delete resize file of writer
-                $this->delete_file($writer->image_path);
-                // $writer->delete // xoa mem
-                $writer->forceDelete(); // xoa han khoi ban ghi
-                return response(json_encode([
-                    "code" => "200",
-                    "value" => "deleted: $writer->name success"
-                ]), 200);
+            if ($writer_id = $this->request->__get("writer_id")) {
+                $writer = $this->writer->find($writer_id);
+                if ($writer && $writer->id) {
+                    // delete resize file of writer
+                    $this->delete_file($writer->image_path);
+                    // $writer->delete // xoa mem
+                    $writer->forceDelete(); // xoa han khoi ban ghi
+                    return response(json_encode([
+                        "code" => "200",
+                        "value" => "deleted: $writer->name success"
+                    ]), 200);
+                }
             }
-        }
         } catch (\Throwable $th) {
             //throw $th;
         }
@@ -111,23 +116,22 @@ class WriterController extends Controller
             $request = $this->request;
             $writer = $this->writer->find($writer_id);
             if ($writer) {
-                if ($file = $request->__get("image_post")){
+                if ($file = $request->__get("image_post")) {
                     $image_upload_path = $this->uploadImage($file, "public/images/writer", "images/resize/writer");
-                    if ($image_upload_path) {
+                    if ($image_upload_path && $writer->image_path) {
                         $this->delete_file($writer->image_path); // xoa file cu de thay bang file moi.
                     }
                 }
                 try {
                     $writer->fill([
-                        "name" => $request->__get("name"),
-                        "email" => $request->__get("email"),
-                        "phone" => $request->__get("phone"),
-                        "address" => $request->__get("address"),
-                        "image_path" => $image_upload_path["file_path"] ? url($image_upload_path["file_path"]) : null,
-                        "name_alias" => $request->__get("alias"),
-                        "active" => $request->__get("active") ?? true,
-                        "date_of_birth" => Carbon::createFromFormat('Y-m-d', $request->__get("date_of_birth")), // date('Y-m-d H:i:s', strtotime($request->__get("date_of_birth")))
-                        "good" => $request->__get("good") ?: null
+                        WriterInterface::ATTR_NAME => $request->__get("name"),
+                        WriterInterface::ATTR_EMAIL => $request->__get("email"),
+                        WriterInterface::ATTR_PHONE => $request->__get("phone"),
+                        WriterInterface::ATTR_ADDRESS => $request->__get("address"),
+                        WriterInterface::ATTR_IMAGE_PATH => $image_upload_path["file_path"] ? url($image_upload_path["file_path"]) : null,
+                        WriterInterface::ATTR_NAME_ALIAS => $request->__get("alias"),
+                        WriterInterface::ATTR_ACTIVE => $request->__get("active") ?? true,
+                        WriterInterface::ATTR_DATE_OF_BIRTH => Carbon::createFromFormat('Y-m-d', $request->__get("date_of_birth")), // date('Y-m-d H:i:s', strtotime($request->__get("date_of_birth")))
                     ]);
                     $result = $writer->save();
                     if ($result) {
@@ -138,7 +142,7 @@ class WriterController extends Controller
                 }
                 return redirect()->back()->with("error", "can not update!");
             }
-        }else {
+        } else {
             return redirect()->back()->with("error", "can not update!");
         }
     }
